Connect edit/publishNow, Add more list config styles
