+++ src/pages/dashboard/configsTab/editor/generalConfig/GeneralConfig.tsx
@@ -3,7 +3,8 @@ import DropDown from '../../../../../components/formFields/DropDown';
 import TextArea from '../../../../../components/formFields/TextArea';
 import TextInput from '../../../../../components/formFields/TextInput';
 import {SurveyConfig} from '../../../../../utilities/types';
-import {AUTH_MODE, DAYS, HOURS, MINUTES, MONTHS, YEARS} from '../constants';
+import {AUTH_MODE} from '../constants';
+import DateSelectorRow from './DateSelectorRow';
 
 interface GeneralConfigProps {
     config: SurveyConfig;
@@ -116,7 +117,7 @@ function GeneralConfig(props: GeneralConfigProps) {
                         <div className='h-12 mr-4 text-xl text-right w-14 leading-12 font-weight-600'>
                             Start:
                         </div>
-                        <DateSelector
+                        <DateSelectorRow
                             date={new Date(props.config.start * 1000)}
                         />
                     </div>
@@ -124,7 +125,7 @@ function GeneralConfig(props: GeneralConfigProps) {
                         <div className='h-12 mr-4 text-xl text-right w-14 leading-12 font-weight-600'>
                             End:
                         </div>
-                        <DateSelector
+                        <DateSelectorRow
                             date={new Date(props.config.end * 1000)}
                         />
                     </div>
@@ -135,62 +136,3 @@ function GeneralConfig(props: GeneralConfigProps) {
 }
 
 export default GeneralConfig;
-
-interface DateSelectorProps {
-    date: any;
-}
-function DateSelector(props: DateSelectorProps) {
-    return (
-        <React.Fragment>
-            <div className='mr-2 w-14'>
-                <DropDown
-                    value={props.date.getDate()}
-                    onChange={() => {}}
-                    options={DAYS}
-                    hideChevron
-                />
-            </div>
-            {DateSeparator('.')}
-            <div className='mx-2 w-36'>
-                <DropDown
-                    value={props.date.getMonth()}
-                    onChange={() => {}}
-                    options={MONTHS}
-                    hideChevron
-                />
-            </div>
-            {DateSeparator('.')}
-            <div className='ml-2 mr-8 w-22'>
-                <DropDown
-                    value={props.date.getYear()}
-                    onChange={() => {}}
-                    options={YEARS}
-                    hideChevron
-                />
-            </div>
-            <div className='mx-2 w-14'>
-                <DropDown
-                    value={props.date.getHours()}
-                    onChange={() => {}}
-                    options={HOURS}
-                    hideChevron
-                />
-            </div>
-            {DateSeparator(':')}
-            <div className='ml-2 w-14'>
-                <DropDown
-                    value={props.date.getMinutes()}
-                    onChange={() => {}}
-                    options={MINUTES}
-                    hideChevron
-                />
-            </div>
-        </React.Fragment>
-    );
-}
-
-const DateSeparator = (char: string) => (
-    <div className='h-12 text-lg text-gray-800 leading-12 font-weight-700'>
-        {char}
-    </div>
-);
